+++ moco-core/src/main/java/com/github/dreamhead/moco/MocoVersion.java
@@ -1,7 +1,7 @@
 package com.github.dreamhead.moco;
 
 public class MocoVersion {
-    public static String VERSION = "0.10.0";
+    public static String VERSION = "0.10.1";
 
     private MocoVersion() {}
 }