+++ moco-doc/ReleaseNotes.md
@@ -1,4 +1,4 @@
-# Release 0.10.1
+# Release 0.10.1 (1-May-2015)
 
 ## APIs
 * Use **httpServer** as API, and **httpserver** has been deprecated, will be removed at next release.